refactor: Define LANGFUSE_ENV_VARS once in shared config
The same three names were listed in langfuse_tracing.ts and inside
ENV_KEYS_TO_SANITIZE in shared/config.ts; adding a variable in one list
but not the other would skip either the presence gate or sanitization.
